+++ lib/src/paint/light_paint_rect.dart
@@ -1,5 +1,4 @@
 import 'dart:math';
-import 'dart:ui';
 
 import 'package:flutter/material.dart';
 import 'package:tutorial_coach_mark/src/target/target_position.dart';
@@ -11,6 +10,7 @@ class LightPaintRect extends CustomPainter {
   final double opacityShadow;
   final double offset;
   final double radius;
+  final BorderSide? borderSide;
 
   LightPaintRect({
     required this.progress,
@@ -19,6 +19,7 @@ class LightPaintRect extends CustomPainter {
     this.opacityShadow = 0.8,
     this.offset = 10,
     this.radius = 10,
+    this.borderSide,
   }) : assert(opacityShadow >= 0 && opacityShadow <= 1);
 
   static Path _drawRectHole(
@@ -87,12 +88,68 @@ class LightPaintRect extends CustomPainter {
       ..close();
   }
 
+  static Path _drawJustHole(
+    Size canvasSize,
+    double x,
+    double y,
+    double w,
+    double h,
+  ) {
+    return Path()
+      ..moveTo(x + w, y)
+      ..lineTo(x + w, y + h)
+      ..lineTo(x, y + h)
+      ..lineTo(x, y)
+      ..close();
+  }
+
+  static Path _drawJustRHole(
+    Size canvasSize,
+    double x,
+    double y,
+    double w,
+    double h,
+    double radius,
+  ) {
+    double diameter = radius * 2;
+
+    return Path()
+      ..moveTo(x, y + radius)
+      ..arcTo(
+        Rect.fromLTWH(x, y, diameter, diameter),
+        pi,
+        pi / 2,
+        false,
+      )
+      ..arcTo(
+        Rect.fromLTWH(x + w - diameter, y, diameter, diameter),
+        3 * pi / 2,
+        pi / 2,
+        false,
+      )
+      ..arcTo(
+        Rect.fromLTWH(x + w - diameter, y + h - diameter, diameter, diameter),
+        0,
+        pi / 2,
+        false,
+      )
+      ..arcTo(
+        Rect.fromLTWH(x, y + h - diameter, diameter, diameter),
+        pi / 2,
+        pi / 2,
+        false,
+      )
+      ..lineTo(x, y + radius)
+      ..close();
+  }
+
   @override
   void paint(Canvas canvas, Size size) {
     if (target.offset == Offset.zero) return;
 
     var maxSize = max(size.width, size.height) +
-        max(target.size.width, target.size.height);
+        max(target.size.width, target.size.height) +
+        target.getBiggerSpaceBorder(size);
 
     double x = -maxSize / 2 * (1 - progress) + target.offset.dx - offset / 2;
 
@@ -108,8 +165,20 @@ class LightPaintRect extends CustomPainter {
           : _drawRectHole(size, x, y, w, h),
       Paint()
         ..style = PaintingStyle.fill
-        ..color = colorShadow.withOpacity(opacityShadow),
+        ..color = colorShadow.withOpacity(opacityShadow)
+        ..strokeWidth = 4,
     );
+    if (borderSide != null && borderSide?.style != BorderStyle.none) {
+      canvas.drawPath(
+        radius > 0
+            ? _drawJustRHole(size, x, y, w, h, radius)
+            : _drawJustHole(size, x, y, w, h),
+        Paint()
+          ..style = PaintingStyle.stroke
+          ..color = borderSide!.color
+          ..strokeWidth = borderSide!.width,
+      );
+    }
   }
 
   @override