+++ lib/src/target/target_content.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
 import '../../tutorial_coach_mark.dart';
@@ -6,12 +7,15 @@ class CustomTargetContentPosition {
   CustomTargetContentPosition({
     this.top,
     this.left,
+    this.right,
     this.bottom,
   });
-  final double? top, left, bottom;
+
+  final double? top, left, right, bottom;
+
   @override
   String toString() {
-    return 'CustomTargetPosition{top: $top, left: $left, bottom: $bottom}';
+    return 'CustomTargetPosition{top: $top, left: $left, right: $right, bottom: $bottom}';
   }
 }
 
@@ -25,15 +29,18 @@ typedef TargetContentBuilder = Widget Function(
 class TargetContent {
   TargetContent({
     this.align = ContentAlign.bottom,
+    this.padding = const EdgeInsets.all(20.0),
     this.child,
     this.customPosition,
     this.builder,
   }) : assert(!(align == ContentAlign.custom && customPosition == null));
 
   final ContentAlign align;
+  final EdgeInsets padding;
   final CustomTargetContentPosition? customPosition;
   final Widget? child;
   final TargetContentBuilder? builder;
+
   @override
   String toString() {
     return 'ContentTarget{align: $align, child: $child}';