+++ lib/tutorial_coach_mark.dart
@@ -1,5 +1,7 @@
 library tutorial_coach_mark;
 
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:tutorial_coach_mark/src/target/target_focus.dart';
 import 'package:tutorial_coach_mark/src/widgets/tutorial_coach_mark_widget.dart';
@@ -12,8 +14,9 @@ export 'package:tutorial_coach_mark/src/util.dart';
 class TutorialCoachMark {
   final BuildContext _context;
   final List<TargetFocus> targets;
-  final Function(TargetFocus)? onClickTarget;
-  final Function(TargetFocus)? onClickOverlay;
+  final FutureOr<void> Function(TargetFocus)? onClickTarget;
+  final FutureOr<void> Function(TargetFocus, TapDownDetails)? onClickTargetWithTapPosition;
+  final FutureOr<void> Function(TargetFocus)? onClickOverlay;
   final Function()? onFinish;
   final double paddingFocus;
   final Function()? onSkip;
@@ -25,27 +28,32 @@ class TutorialCoachMark {
   final double opacityShadow;
   final GlobalKey<TutorialCoachMarkWidgetState> _widgetKey = GlobalKey();
   final Duration focusAnimationDuration;
+  final Duration unFocusAnimationDuration;
   final Duration pulseAnimationDuration;
+  final bool pulseEnable;
   final Widget? skipWidget;
 
   OverlayEntry? _overlayEntry;
 
   TutorialCoachMark(this._context,
       {required this.targets,
-      this.colorShadow = Colors.black,
-      this.onClickTarget,
-      this.onClickOverlay,
-      this.onFinish,
-      this.paddingFocus = 10,
-      this.onSkip,
-      this.alignSkip = Alignment.bottomRight,
-      this.textSkip = "SKIP",
-      this.textStyleSkip = const TextStyle(color: Colors.white),
-      this.hideSkip = false,
-      this.opacityShadow = 0.8,
-      this.focusAnimationDuration = const Duration(milliseconds: 600),
-      this.pulseAnimationDuration = const Duration(milliseconds: 500),
-      this.skipWidget})
+        this.colorShadow = Colors.black,
+        this.onClickTarget,
+        this.onClickTargetWithTapPosition,
+        this.onClickOverlay,
+        this.onFinish,
+        this.paddingFocus = 10,
+        this.onSkip,
+        this.alignSkip = Alignment.bottomRight,
+        this.textSkip = "SKIP",
+        this.textStyleSkip = const TextStyle(color: Colors.white),
+        this.hideSkip = false,
+        this.opacityShadow = 0.8,
+        this.focusAnimationDuration = const Duration(milliseconds: 600),
+        this.unFocusAnimationDuration = const Duration(milliseconds: 600),
+        this.pulseAnimationDuration = const Duration(milliseconds: 500),
+        this.pulseEnable = true,
+        this.skipWidget})
       : assert(opacityShadow >= 0 && opacityShadow <= 1);
 
   OverlayEntry _buildOverlay() {
@@ -55,6 +63,7 @@ class TutorialCoachMark {
           key: _widgetKey,
           targets: targets,
           clickTarget: onClickTarget,
+          onClickTargetWithTapPosition: onClickTargetWithTapPosition,
           clickOverlay: onClickOverlay,
           paddingFocus: paddingFocus,
           onClickSkip: skip,
@@ -66,18 +75,20 @@ class TutorialCoachMark {
           colorShadow: colorShadow,
           opacityShadow: opacityShadow,
           focusAnimationDuration: focusAnimationDuration,
+          unFocusAnimationDuration: unFocusAnimationDuration,
           pulseAnimationDuration: pulseAnimationDuration,
+          pulseEnable: pulseEnable,
           finish: finish,
         );
       },
     );
   }
 
-  void show() {
+  void show({bool rootOverlay = false}) {
     Future.delayed(Duration.zero, () {
       if (_overlayEntry == null) {
         _overlayEntry = _buildOverlay();
-        Overlay.of(_context)!.insert(_overlayEntry!);
+        Overlay.of(_context, rootOverlay: rootOverlay)?.insert(_overlayEntry!);
       }
     });
   }