+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+# 1.2.3
+- Fix warn build in Flutter 3.0.
+- Collect target tap position. [PR #116](https://github.com/RafaelBarbosatec/tutorial_coach_mark/pull/116)
+
+# 1.2.2
+- Add unfocus animation duration. [PR #111](https://github.com/RafaelBarbosatec/tutorial_coach_mark/pull/111)
+- Update Build gradle, kotlin, fix manifest. [PR #110](https://github.com/RafaelBarbosatec/tutorial_coach_mark/pull/110)
+
+# 1.2.1
+- Add `BorderSide` param in `TargetFocus`. [PR #96](https://github.com/RafaelBarbosatec/tutorial_coach_mark/pull/96)
+- Added right property for custom alignment. [PR #101](https://github.com/RafaelBarbosatec/tutorial_coach_mark/pull/101)
+
+# 1.2.0
+- Add `pulseEnable` param in `TutorialCoachMark` to enable/disable pulse animation. [PR #88](https://github.com/RafaelBarbosatec/tutorial_coach_mark/pull/88)
+- Add `padding` param in `TargetContent` to customize target padding. [PR #84](https://github.com/RafaelBarbosatec/tutorial_coach_mark/pull/84)
+- Add optional param `rootOverlay` in `TutorialCoachMark.show()`. Suggestion date in issue [93](https://github.com/RafaelBarbosatec/tutorial_coach_mark/issues/93)
+- The methods `onClickTarget` and `onClickOverlay` now they can be asynchronous and the next target will only be called after processing something added to it.
+
+# 1.1.1
+- Fiz error [issue 82](https://github.com/RafaelBarbosatec/tutorial_coach_mark/issues/82)
+
+# 1.1.0
+- Add web support! [PR #79](https://github.com/RafaelBarbosatec/tutorial_coach_mark/pull/79) Thanks [Will Reppun
+](https://github.com/wreppun)
+
 # 1.0.3
 - Fix blink when start tutorial with shape `ShapeLightFocus.RRec`